+++ common/Curves/CircleCurve.cs
@@ -40,9 +40,9 @@ public class CircleCurve(Vector2 startPoint, Vector2 midPoint, Vector2 endPoint)
         length = Math.Abs((endAngle - startAngle) * radius);
         var precision = (int)(length / 4);
 
-        for (var i = 1; i < length; ++i)
+        for (var i = 1f; i < length; ++i)
         {
-            var progress = (float)i / precision;
+            var progress = i / precision;
             var angle = endAngle * progress + startAngle * (1 - progress);
 
             distancePosition.Add((progress * length, new Vector2(MathF.Cos(angle) * radius, MathF.Sin(angle) * radius) + centre));
